+++ apps/webapp/app/routes/api.v1.runs.$runId.tasks.$id.callback.$secret.ts
@@ -1,3 +1,3 @@
-import { action } from "./api.v1.tasks.$id.callback.$secret";
+import { action } from "./api.v1.tasks.$id.callback.$secret/route";
 
 export { action };
